# frontend/openapi.yaml
openapi: 3.0.1
info:
  title: REST API LLS
  description: LLS - Lehrlings Liefer System | Apprentice Delivery System
  contact:
    name: Lars-Thorsten Bilek
    email: lars.bilek@gmx.at
  version: "1.0"
servers:
- url: https://localhost:9002
  description: Generated server url
security:
- bearerAuth: []
...
paths:
...
  /api/portal/grocery-working-day/:
    get:
      tags:
      - AdminModGroceryWorkingDayManagementResource
      description: load all grocery working days with pagination and filter
      operationId: loadAllGroceryWorkingDays
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: goingUserId
        in: query
        required: false
        schema:
          type: string
      - name: payingUserId
        in: query
        required: false
        schema:
          type: string
      - name: date
        in: query
        required: false
        schema:
          type: string
          format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWrapperGroceryWorkingDayDetailDTO'
    put:
      tags:
      - AdminModGroceryWorkingDayManagementResource
      description: update grocery working day
      operationId: updateGroceryWorkingDay
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroceryWorkingDayDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroceryWorkingDayDetailDTO'
    post:
      tags:
      - AdminModGroceryWorkingDayManagementResource
      description: create new grocery working day
      operationId: createGroceryWorkingDay
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroceryWorkingDayDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroceryWorkingDayDetailDTO'
...
  /api/order/today/orders:
    get:
      tags:
      - UserOrderResource
      description: load all order for the paying user with pagination and filter
      operationId: loadTodayOrderForPayingUser
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderDetailDTO'
    post:
      tags:
      - UserOrderResource
      description: create order for the current customer
      operationId: createOrderForCurrentCustomer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailDTO'
...
  /api/portal/order/{id}:
    get:
      tags:
      - AdminModOrderManagementResource
      description: load order by id
      operationId: loadOrderById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailDTO'
  /api/portal/order/:
    get:
      tags:
      - AdminModOrderManagementResource
      description: load all orders with pagination and filter
      operationId: loadAllOrders
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: searchTerm
        in: query
        required: false
        schema:
          type: string
      - name: isPayed
        in: query
        required: false
        schema:
          type: boolean
      - name: paymentType
        in: query
        required: false
        schema:
          type: string
          enum:
          - CASH
          - PAYPAL
          - CREDIT_CARD
          - BANK_TRANSFER
      - name: orderDateFrom
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: orderDateTo
        in: query
        required: false
        schema:
          type: string
          format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWrapperOrderOverviewDTO'
...
  /api/order/products:
    get:
      tags:
      - UserOrderResource
      description: load all products for shop
      operationId: loadAllProductsForShop
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: searchTerm
        in: query
        required: false
        schema:
          type: string
      - name: categoryId
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWrapperProductDetailDTO'
  /api/order/paying-user/orders:
    get:
      tags:
      - UserOrderResource
      description: load all order for the paying user with pagination and filter
      operationId: loadAllOrdersForPayingUser
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWrapperOrderDetailDTO'
...
components:
  schemas:
...
    UserDetailDTO:
      type: object
      properties:
        id:
          type: string
        version:
          type: integer
          format: int32
        firstname:
          type: string
        lastname:
          type: string
        status:
          type: string
        phoneNumber:
          type: string
        location:
          type: string
        iban:
          type: string
        paypalLink:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
            enum:
            - ADMIN
            - MOD
            - USER
        enabled:
          type: boolean
...
    CategoryDetailDTO:
      type: object
      properties:
        id:
          type: string
        version:
          type: integer
          format: int32
        name:
          type: string
        description:
          type: string
    ProductDetailDTO:
      type: object
      properties:
        id:
          type: string
        version:
          type: integer
          format: int32
        name:
          type: string
        producer:
          type: string
        quantity:
          type: string
        price:
          type: number
          format: double
        isActive:
          type: boolean
        isChecked:
          type: boolean
        category:
          $ref: '#/components/schemas/CategoryDetailDTO'
    UpdateGroceryWorkingDayDTO:
      required:
      - id
      - status
      type: object
      properties:
        id:
          type: string
        version:
          minimum: 0
          type: integer
          format: int32
        goingUserIds:
          type: array
          items:
            type: string
        payingUserId:
          type: string
        status:
          type: string
          enum:
          - OPEN
          - CLOSED
    GroceryWorkingDayDetailDTO:
      type: object
      properties:
        id:
          type: string
        version:
          type: integer
          format: int32
        date:
          type: string
          format: date
        goingUsers:
          type: array
          items:
            $ref: '#/components/schemas/UserOverviewDTO'
        payingUser:
          $ref: '#/components/schemas/UserOverviewDTO'
        status:
          type: string
          enum:
          - OPEN
          - CLOSED
    UserOverviewDTO:
      type: object
      properties:
        id:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
...
    CreateGroceryWorkingDayDTO:
      required:
      - date
      type: object
      properties:
        date:
          type: string
          format: date
        goingUserIds:
          type: array
          items:
            type: string
        payingUserId:
          type: string
...
    CreateOrderDTO:
      required:
      - paymentType
      type: object
      properties:
        paymentType:
          type: string
          enum:
          - CASH
          - PAYPAL
          - CREDIT_CARD
          - BANK_TRANSFER
        orderedItems:
          maxItems: 2147483647
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/CreateOrderOrderItemDTO'
    CreateOrderOrderItemDTO:
      required:
      - productId
      type: object
      properties:
        productId:
          type: string
        orderedQuantity:
          maximum: 2147483647
          minimum: 1
          type: integer
          format: int32
        note:
          type: string
    OrderDetailDTO:
      type: object
      properties:
        id:
          type: string
        user:
          $ref: '#/components/schemas/UserDetailDTO'
        isPayed:
          type: boolean
        groceryWorkingDay:
          $ref: '#/components/schemas/GroceryWorkingDayDetailDTO'
        paymentType:
          type: string
          enum:
          - CASH
          - PAYPAL
          - CREDIT_CARD
          - BANK_TRANSFER
        orderItems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemDetailDTO'
        status:
          type: string
          enum:
          - ORDERED
          - IN_PROGRESS
          - READY_TO_PAY
          - DONE
          - CANCELLED
    OrderItemDetailDTO:
      type: object
      properties:
        id:
          type: string
        orderedQuantity:
          type: integer
          format: int32
        note:
          type: string
        oldProductPrice:
          type: number
          format: double
        product:
          $ref: '#/components/schemas/ProductDetailDTO'
...
    OrderItemOverviewDTO:
      type: object
      properties:
        id:
          type: string
        orderedQuantity:
          type: integer
          format: int32
        productName:
          type: string
        oldProductPrice:
          type: number
          format: double
    OrderOverviewDTO:
      type: object
      properties:
        id:
          type: string
        user:
          $ref: '#/components/schemas/UserOverviewDTO'
        isPayed:
          type: boolean
        orderDate:
          type: string
          format: date
        paymentType:
          type: string
          enum:
          - CASH
          - PAYPAL
          - CREDIT_CARD
          - BANK_TRANSFER
        orderItems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemOverviewDTO'
        status:
          type: string
          enum:
          - ORDERED
          - IN_PROGRESS
          - READY_TO_PAY
          - DONE
          - CANCELLED
    ResponseWrapperOrderOverviewDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/OrderOverviewDTO'
        totalElements:
          type: integer
          format: int64
    ResponseWrapperGroceryWorkingDayDetailDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/GroceryWorkingDayDetailDTO'
        totalElements:
          type: integer
          format: int64
...
    ResponseWrapperProductDetailDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProductDetailDTO'
        totalElements:
          type: integer
          format: int64
    ResponseWrapperOrderDetailDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/OrderDetailDTO'
        totalElements:
          type: integer
          format: int64
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT